+++ projects/ngneat/transloco-utils/src/lib/types.ts
@@ -1,6 +1,7 @@
 export interface TranslocoConfig {
   rootTranslationsPath?: string;
   defaultLang?: string;
+  scopedLibs?: string[];
   scopePathMap?: { [key: string]: string };
   langs?: string[];
   keysManager?: {