+++ projects/ngneat/transloco/src/lib/get-fallbacks-loaders.ts
@@ -1,17 +1,18 @@
 import { from } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { resolveLoader } from './resolve-loader';
-import { TranslocoLoader } from './transloco.loader';
+import { TranslocoLoader, TranslocoLoaderData } from './transloco.loader';
 import { InlineLoader } from './types';
 
 export function getFallbacksLoaders(
   mainPath: string,
   fallbackPath: string,
   mainLoader: TranslocoLoader,
-  inlineLoader: InlineLoader
+  inlineLoader: InlineLoader,
+  data: TranslocoLoaderData
 ) {
   return [mainPath, fallbackPath].map(path => {
-    const loader = resolveLoader(path, mainLoader, inlineLoader);
+    const loader = resolveLoader(path, mainLoader, inlineLoader, data);
 
     return from(loader).pipe(
       map(translation => ({