+++ schematics/src/utils/translations.ts
@@ -12,9 +12,9 @@ import { TranslationFileTypes } from '../ng-add/schema';
 const p = require('path');
 const fs = require('fs');
 
-export function jsonTranslationFileCreator(source, lang) {
+export function jsonTranslationFileCreator(source, lang, path) {
   return source.create(
-    `${lang}.json`,
+    p.join(path, `${lang}.json`),
     `{
   "title": "transloco ${lang}",
   "dynamic": "transloco {{value}}"
@@ -23,9 +23,9 @@ export function jsonTranslationFileCreator(source, lang) {
   );
 }
 
-export function typescriptTranslationFileCreator(source, lang) {
+export function typescriptTranslationFileCreator(source, lang, path) {
   return source.create(
-    `${lang}.ts`,
+    p.join(path, `${lang}.ts`),
     `export default {
   title: "transloco ${lang}",
   dynamic: "transloco {{value}}"
@@ -51,7 +51,7 @@ export function createTranslateFilesFromOptions(
   host: Tree,
   options: { translateType?: TranslationFileTypes; langs: string[] },
   translationFilePath
-): Source {
+): Tree {
   const extension = options.translateType === TranslationFileTypes.Typescript ? '.ts' : '.json';
   const translationCreator =
     options.translateType === TranslationFileTypes.Typescript
@@ -60,13 +60,13 @@ export function createTranslateFilesFromOptions(
 
   checkIfTranslationFilesExist(translationFilePath, options.langs, extension);
 
-  return apply(source(createTranslateFiles(options.langs, translationCreator)), [move('/', translationFilePath)]);
+  return createTranslateFiles(options.langs, translationCreator, translationFilePath);
 }
 
-export function createTranslateFiles(langs: string[], creator): HostTree {
+export function createTranslateFiles(langs: string[], creator, path): HostTree {
   const treeSource = new EmptyTree();
   langs.forEach(lang => {
-    creator(treeSource, lang);
+    creator(treeSource, lang, path);
   });
 
   return treeSource;