+++ src/app/loaders/http.loader.ts
@@ -6,7 +6,7 @@ import { Injectable } from '@angular/core';
 export class HttpLoader implements TranslocoLoader {
   constructor(private http: HttpClient) {}
 
-  getTranslation(langPath: string) {
+  getTranslation(langPath: string, { scope }) {
     return this.http.get<Translation>(`/assets/i18n/${langPath}.json`);
   }
 }